refactor(test, aten): address PR #49 review feedback
Two reviewer threads on PR #49, addressed together:

Tests — align v2v test suites with docs/TEST_GUIDE.md Minimal Test Template:
- move test/python/test_*_v2v.py to test/rbln/ (the new dir was unused)
- free `def test_*` -> `TestCase` classes (TestCatV2V, TestIndexSelectV2V,
  TestCopyV2V, TestIndexCopyV2V, TestRepeatInterleaveV2V)
- `@pytest.mark.parametrize` -> `@parametrize` from torch internal; dtype
  params -> `@dtypes(...)`; module ends with
  `instantiate_device_type_tests(..., only_for="privateuse1")` + `run_tests()`
- add `@pytest.mark.test_set_ci` + `@pytest.mark.usefixtures("enable_deploy_mode")`
  to each class; drop the import-time `os.environ.setdefault` from utils_v2v.py
- remove the now-subsumed `TestRegisteredNativeOps.test_cat` smoke case

V2VBatch — flip "RAII flush by default" to "explicit submit() required":
- destructor is now a leak-prevention safety net only: warns when entries
  are pending on a normal path (guarded by std::uncaught_exceptions() == 0)
  and never issues backend calls, so a rebel rejection during stack unwind
  can no longer terminate the process
- add explicit batch.submit() at all 4 call sites (RBLNStridedV2V.cpp,
  RBLNTensorShape.cpp, RBLNTensorAdvancedIndexing.cpp x 2); drop the
  implicit-flush comments those sites used to need
- replace DestructorFlush gtest with DestructorDoesNotFlush, and add
  DestructorIsNoexceptDuringUnwind covering the unwind path
